Struct relay_auth::SecretKey

pub struct SecretKey { /* private fields */ }

Represents the secret key of an Relay.

Secret keys are based on ed25519 but this should be considered an implementation detail for now. We only ever represent public keys on the wire as opaque ascii encoded strings of arbitrary format or length.

Implementations

impl SecretKey

pub fn sign(&self, data: &[u8]) -> String

Signs some data with the secret key and returns the signature.

This is will sign with the default header.

pub fn sign_with_header(&self, data: &[u8], header: &SignatureHeader) -> String

Signs some data with the secret key and a specific header and then returns the signature.

The default behavior is to attach the timestamp in the header to the signature so that old signatures on verification can be rejected.

pub fn pack<S: Serialize>(&self, data: S) -> (Vec<u8>, String)

Packs some serializable data into JSON and signs it with the default header.

pub fn pack_with_header<S: Serialize>( &self, data: S, header: &SignatureHeader ) -> (Vec<u8>, String)

Packs some serializable data into JSON and signs it with the specified header.