Crate relay_auth

Structs

ParseRelayVersionError

PublicKey

Represents the public key of an relay.

RegisterChallenge

Represents the response the server is supposed to send to a register request.

RegisterRequest

Represents a request for registration with the upstream.

RegisterResponse

Represents a response to a register challenge.

RegisterState

A state structure containing relevant information from RegisterRequest.

Registration

Represents the final registration.

RelayVersion

The version of a Relay.

SecretKey

Represents the secret key of an relay.

SignatureHeader

A wrapper around packed data that adds a timestamp.

SignedRegisterState

An encoded and signed RegisterState.

Enums

KeyParseError

Raised if a key could not be parsed.

UnpackError

Raised to indicate failure on unpacking.

Constants

VERSION_MAJOR

VERSION_MINOR

VERSION_PATCH

Functions

generate_key_pair

Generates a secret + public key pair.

generate_relay_id

Generates an relay ID.

Type Definitions

RelayId

Alias for relay IDs (UUIDs).